+++ script/tool/test/pubspec_check_command_test.dart
@@ -12,6 +12,88 @@ import 'package:test/test.dart';
 import 'mocks.dart';
 import 'util.dart';
 
+/// Returns the top section of a pubspec.yaml for a package named [name],
+/// for either a flutter/packages or flutter/plugins package depending on
+/// the values of [isPlugin].
+///
+/// By default it will create a header that includes all of the expected
+/// values, elements can be changed via arguments to create incorrect
+/// entries.
+///
+/// If [includeRepository] is true, by default the path in the link will
+/// be "packages/[name]"; a different "packages"-relative path can be
+/// provided with [repositoryPackagesDirRelativePath].
+String _headerSection(
+  String name, {
+  bool isPlugin = false,
+  bool includeRepository = true,
+  String? repositoryPackagesDirRelativePath,
+  bool includeHomepage = false,
+  bool includeIssueTracker = true,
+  bool publishable = true,
+}) {
+  final String repositoryPath = repositoryPackagesDirRelativePath ?? name;
+  final String repoLink = 'https://github.com/flutter/'
+      '${isPlugin ? 'plugins' : 'packages'}/tree/master/'
+      'packages/$repositoryPath';
+  final String issueTrackerLink = 'https://github.com/flutter/flutter/issues?'
+      'q=is%3Aissue+is%3Aopen+label%3A%22p%3A+$name%22';
+  return '''
+name: $name
+${includeRepository ? 'repository: $repoLink' : ''}
+${includeHomepage ? 'homepage: $repoLink' : ''}
+${includeIssueTracker ? 'issue_tracker: $issueTrackerLink' : ''}
+version: 1.0.0
+${publishable ? '' : 'publish_to: \'none\''}
+''';
+}
+
+String _environmentSection() {
+  return '''
+environment:
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=2.0.0"
+''';
+}
+
+String _flutterSection({
+  bool isPlugin = false,
+  String? implementedPackage,
+}) {
+  final String pluginEntry = '''
+  plugin:
+${implementedPackage == null ? '' : '    implements: $implementedPackage'}
+    platforms:
+''';
+  return '''
+flutter:
+${isPlugin ? pluginEntry : ''}
+''';
+}
+
+String _dependenciesSection() {
+  return '''
+dependencies:
+  flutter:
+    sdk: flutter
+''';
+}
+
+String _devDependenciesSection() {
+  return '''
+dev_dependencies:
+  flutter_test:
+    sdk: flutter
+''';
+}
+
+String _falseSecretsSection() {
+  return '''
+false_secrets:
+  - /lib/main.dart
+''';
+}
+
 void main() {
   group('test pubspec_check_command', () {
     late CommandRunner<void> runner;
@@ -37,99 +119,16 @@ void main() {
       runner.addCommand(command);
     });
 
-    /// Returns the top section of a pubspec.yaml for a package named [name],
-    /// for either a flutter/packages or flutter/plugins package depending on
-    /// the values of [isPlugin].
-    ///
-    /// By default it will create a header that includes all of the expected
-    /// values, elements can be changed via arguments to create incorrect
-    /// entries.
-    ///
-    /// If [includeRepository] is true, by default the path in the link will
-    /// be "packages/[name]"; a different "packages"-relative path can be
-    /// provided with [repositoryPackagesDirRelativePath].
-    String headerSection(
-      String name, {
-      bool isPlugin = false,
-      bool includeRepository = true,
-      String? repositoryPackagesDirRelativePath,
-      bool includeHomepage = false,
-      bool includeIssueTracker = true,
-      bool publishable = true,
-    }) {
-      final String repositoryPath = repositoryPackagesDirRelativePath ?? name;
-      final String repoLink = 'https://github.com/flutter/'
-          '${isPlugin ? 'plugins' : 'packages'}/tree/master/'
-          'packages/$repositoryPath';
-      final String issueTrackerLink =
-          'https://github.com/flutter/flutter/issues?'
-          'q=is%3Aissue+is%3Aopen+label%3A%22p%3A+$name%22';
-      return '''
-name: $name
-${includeRepository ? 'repository: $repoLink' : ''}
-${includeHomepage ? 'homepage: $repoLink' : ''}
-${includeIssueTracker ? 'issue_tracker: $issueTrackerLink' : ''}
-version: 1.0.0
-${publishable ? '' : 'publish_to: \'none\''}
-''';
-    }
-
-    String environmentSection() {
-      return '''
-environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=2.0.0"
-''';
-    }
-
-    String flutterSection({
-      bool isPlugin = false,
-      String? implementedPackage,
-    }) {
-      final String pluginEntry = '''
-  plugin:
-${implementedPackage == null ? '' : '    implements: $implementedPackage'}
-    platforms:
-''';
-      return '''
-flutter:
-${isPlugin ? pluginEntry : ''}
-''';
-    }
-
-    String dependenciesSection() {
-      return '''
-dependencies:
-  flutter:
-    sdk: flutter
-''';
-    }
-
-    String devDependenciesSection() {
-      return '''
-dev_dependencies:
-  flutter_test:
-    sdk: flutter
-''';
-    }
-
-    String falseSecretsSection() {
-      return '''
-false_secrets:
-  - /lib/main.dart
-''';
-    }
-
     test('passes for a plugin following conventions', () async {
       final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
-${headerSection('plugin', isPlugin: true)}
-${environmentSection()}
-${flutterSection(isPlugin: true)}
-${dependenciesSection()}
-${devDependenciesSection()}
-${falseSecretsSection()}
+${_headerSection('plugin', isPlugin: true)}
+${_environmentSection()}
+${_flutterSection(isPlugin: true)}
+${_dependenciesSection()}
+${_devDependenciesSection()}
+${_falseSecretsSection()}
 ''');
 
       final List<String> output = await runCapturingPrint(runner, <String>[
@@ -150,12 +149,12 @@ ${falseSecretsSection()}
       final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
-${headerSection('plugin')}
-${environmentSection()}
-${dependenciesSection()}
-${devDependenciesSection()}
-${flutterSection()}
-${falseSecretsSection()}
+${_headerSection('plugin')}
+${_environmentSection()}
+${_dependenciesSection()}
+${_devDependenciesSection()}
+${_flutterSection()}
+${_falseSecretsSection()}
 ''');
 
       final List<String> output = await runCapturingPrint(runner, <String>[
@@ -177,9 +176,9 @@ ${falseSecretsSection()}
       packageDirectory.createSync(recursive: true);
 
       packageDirectory.childFile('pubspec.yaml').writeAsStringSync('''
-${headerSection('package')}
-${environmentSection()}
-${dependenciesSection()}
+${_headerSection('package')}
+${_environmentSection()}
+${_dependenciesSection()}
 ''');
 
       final List<String> output = await runCapturingPrint(runner, <String>[
@@ -199,11 +198,11 @@ ${dependenciesSection()}
       final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
-${headerSection('plugin', isPlugin: true, includeHomepage: true)}
-${environmentSection()}
-${flutterSection(isPlugin: true)}
-${dependenciesSection()}
-${devDependenciesSection()}
+${_headerSection('plugin', isPlugin: true, includeHomepage: true)}
+${_environmentSection()}
+${_flutterSection(isPlugin: true)}
+${_dependenciesSection()}
+${_devDependenciesSection()}
 ''');
 
       Error? commandError;
@@ -226,11 +225,11 @@ ${devDependenciesSection()}
       final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
-${headerSection('plugin', isPlugin: true, includeRepository: false)}
-${environmentSection()}
-${flutterSection(isPlugin: true)}
-${dependenciesSection()}
-${devDependenciesSection()}
+${_headerSection('plugin', isPlugin: true, includeRepository: false)}
+${_environmentSection()}
+${_flutterSection(isPlugin: true)}
+${_dependenciesSection()}
+${_devDependenciesSection()}
 ''');
 
       Error? commandError;
@@ -252,11 +251,11 @@ ${devDependenciesSection()}
       final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
-${headerSection('plugin', isPlugin: true, includeHomepage: true, includeRepository: false)}
-${environmentSection()}
-${flutterSection(isPlugin: true)}
-${dependenciesSection()}
-${devDependenciesSection()}
+${_headerSection('plugin', isPlugin: true, includeHomepage: true, includeRepository: false)}
+${_environmentSection()}
+${_flutterSection(isPlugin: true)}
+${_dependenciesSection()}
+${_devDependenciesSection()}
 ''');
 
       Error? commandError;
@@ -279,11 +278,11 @@ ${devDependenciesSection()}
       final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
-${headerSection('plugin', isPlugin: true, repositoryPackagesDirRelativePath: 'different_plugin')}
-${environmentSection()}
-${flutterSection(isPlugin: true)}
-${dependenciesSection()}
-${devDependenciesSection()}
+${_headerSection('plugin', isPlugin: true, repositoryPackagesDirRelativePath: 'different_plugin')}
+${_environmentSection()}
+${_flutterSection(isPlugin: true)}
+${_dependenciesSection()}
+${_devDependenciesSection()}
 ''');
 
       Error? commandError;
@@ -305,11 +304,11 @@ ${devDependenciesSection()}
       final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
-${headerSection('plugin', isPlugin: true, includeIssueTracker: false)}
-${environmentSection()}
-${flutterSection(isPlugin: true)}
-${dependenciesSection()}
-${devDependenciesSection()}
+${_headerSection('plugin', isPlugin: true, includeIssueTracker: false)}
+${_environmentSection()}
+${_flutterSection(isPlugin: true)}
+${_dependenciesSection()}
+${_devDependenciesSection()}
 ''');
 
       Error? commandError;
@@ -331,11 +330,11 @@ ${devDependenciesSection()}
       final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
-${headerSection('plugin', isPlugin: true)}
-${flutterSection(isPlugin: true)}
-${dependenciesSection()}
-${devDependenciesSection()}
-${environmentSection()}
+${_headerSection('plugin', isPlugin: true)}
+${_flutterSection(isPlugin: true)}
+${_dependenciesSection()}
+${_devDependenciesSection()}
+${_environmentSection()}
 ''');
 
       Error? commandError;
@@ -358,11 +357,11 @@ ${environmentSection()}
       final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
-${headerSection('plugin', isPlugin: true)}
-${flutterSection(isPlugin: true)}
-${environmentSection()}
-${dependenciesSection()}
-${devDependenciesSection()}
+${_headerSection('plugin', isPlugin: true)}
+${_flutterSection(isPlugin: true)}
+${_environmentSection()}
+${_dependenciesSection()}
+${_devDependenciesSection()}
 ''');
 
       Error? commandError;
@@ -385,11 +384,11 @@ ${devDependenciesSection()}
       final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
-${headerSection('plugin', isPlugin: true)}
-${environmentSection()}
-${flutterSection(isPlugin: true)}
-${devDependenciesSection()}
-${dependenciesSection()}
+${_headerSection('plugin', isPlugin: true)}
+${_environmentSection()}
+${_flutterSection(isPlugin: true)}
+${_devDependenciesSection()}
+${_dependenciesSection()}
 ''');
 
       Error? commandError;
@@ -412,11 +411,11 @@ ${dependenciesSection()}
       final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
-${headerSection('plugin', isPlugin: true)}
-${environmentSection()}
-${devDependenciesSection()}
-${flutterSection(isPlugin: true)}
-${dependenciesSection()}
+${_headerSection('plugin', isPlugin: true)}
+${_environmentSection()}
+${_devDependenciesSection()}
+${_flutterSection(isPlugin: true)}
+${_dependenciesSection()}
 ''');
 
       Error? commandError;
@@ -439,12 +438,12 @@ ${dependenciesSection()}
       final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
-${headerSection('plugin', isPlugin: true)}
-${environmentSection()}
-${flutterSection(isPlugin: true)}
-${dependenciesSection()}
-${falseSecretsSection()}
-${devDependenciesSection()}
+${_headerSection('plugin', isPlugin: true)}
+${_environmentSection()}
+${_flutterSection(isPlugin: true)}
+${_dependenciesSection()}
+${_falseSecretsSection()}
+${_devDependenciesSection()}
 ''');
 
       Error? commandError;
@@ -469,11 +468,11 @@ ${devDependenciesSection()}
           'plugin_a_foo', packagesDir.childDirectory('plugin_a'));
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
-${headerSection('plugin_a_foo', isPlugin: true)}
-${environmentSection()}
-${flutterSection(isPlugin: true)}
-${dependenciesSection()}
-${devDependenciesSection()}
+${_headerSection('plugin_a_foo', isPlugin: true)}
+${_environmentSection()}
+${_flutterSection(isPlugin: true)}
+${_dependenciesSection()}
+${_devDependenciesSection()}
 ''');
 
       Error? commandError;
@@ -497,11 +496,11 @@ ${devDependenciesSection()}
           'plugin_a_foo', packagesDir.childDirectory('plugin_a'));
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
-${headerSection('plugin_a_foo', isPlugin: true)}
-${environmentSection()}
-${flutterSection(isPlugin: true, implementedPackage: 'plugin_a_foo')}
-${dependenciesSection()}
-${devDependenciesSection()}
+${_headerSection('plugin_a_foo', isPlugin: true)}
+${_environmentSection()}
+${_flutterSection(isPlugin: true, implementedPackage: 'plugin_a_foo')}
+${_dependenciesSection()}
+${_devDependenciesSection()}
 ''');
 
       Error? commandError;
@@ -525,15 +524,15 @@ ${devDependenciesSection()}
           'plugin_a_foo', packagesDir.childDirectory('plugin_a'));
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
-${headerSection(
+${_headerSection(
         'plugin_a_foo',
         isPlugin: true,
         repositoryPackagesDirRelativePath: 'plugin_a/plugin_a_foo',
       )}
-${environmentSection()}
-${flutterSection(isPlugin: true, implementedPackage: 'plugin_a')}
-${dependenciesSection()}
-${devDependenciesSection()}
+${_environmentSection()}
+${_flutterSection(isPlugin: true, implementedPackage: 'plugin_a')}
+${_dependenciesSection()}
+${_devDependenciesSection()}
 ''');
 
       final List<String> output =
@@ -553,15 +552,15 @@ ${devDependenciesSection()}
           createFakePlugin('plugin_a', packagesDir.childDirectory('plugin_a'));
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
-${headerSection(
+${_headerSection(
         'plugin_a',
         isPlugin: true,
         repositoryPackagesDirRelativePath: 'plugin_a/plugin_a',
       )}
-${environmentSection()}
-${flutterSection(isPlugin: true)}
-${dependenciesSection()}
-${devDependenciesSection()}
+${_environmentSection()}
+${_flutterSection(isPlugin: true)}
+${_dependenciesSection()}
+${_devDependenciesSection()}
 ''');
 
       final List<String> output =
@@ -583,16 +582,16 @@ ${devDependenciesSection()}
           packagesDir.childDirectory('plugin_a'));
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
-${headerSection(
+${_headerSection(
         'plugin_a_platform_interface',
         isPlugin: true,
         repositoryPackagesDirRelativePath:
             'plugin_a/plugin_a_platform_interface',
       )}
-${environmentSection()}
-${flutterSection(isPlugin: true)}
-${dependenciesSection()}
-${devDependenciesSection()}
+${_environmentSection()}
+${_flutterSection(isPlugin: true)}
+${_dependenciesSection()}
+${_devDependenciesSection()}
 ''');
 
       final List<String> output =
@@ -614,11 +613,11 @@ ${devDependenciesSection()}
       // Environment section is in the wrong location.
       // Missing 'implements'.
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
-${headerSection('plugin_a_foo', isPlugin: true, publishable: false)}
-${flutterSection(isPlugin: true)}
-${dependenciesSection()}
-${devDependenciesSection()}
-${environmentSection()}
+${_headerSection('plugin_a_foo', isPlugin: true, publishable: false)}
+${_flutterSection(isPlugin: true)}
+${_dependenciesSection()}
+${_devDependenciesSection()}
+${_environmentSection()}
 ''');
 
       Error? commandError;
@@ -644,17 +643,17 @@ ${environmentSection()}
       // Missing metadata that is only useful for published packages, such as
       // repository and issue tracker.
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
-${headerSection(
+${_headerSection(
         'plugin',
         isPlugin: true,
         publishable: false,
         includeRepository: false,
         includeIssueTracker: false,
       )}
-${environmentSection()}
-${flutterSection(isPlugin: true)}
-${dependenciesSection()}
-${devDependenciesSection()}
+${_environmentSection()}
+${_flutterSection(isPlugin: true)}
+${_dependenciesSection()}
+${_devDependenciesSection()}
 ''');
 
       final List<String> output =
@@ -664,6 +663,53 @@ ${devDependenciesSection()}
         output,
         containsAllInOrder(<Matcher>[
           contains('Running for plugin...'),
+          contains('No issues found!'),
+        ]),
+      );
+    });
+  });
+
+  group('test pubspec_check_command on Windows', () {
+    late CommandRunner<void> runner;
+    late RecordingProcessRunner processRunner;
+    late FileSystem fileSystem;
+    late MockPlatform mockPlatform;
+    late Directory packagesDir;
+
+    setUp(() {
+      fileSystem = MemoryFileSystem(style: FileSystemStyle.windows);
+      mockPlatform = MockPlatform(isWindows: true);
+      packagesDir = fileSystem.currentDirectory.childDirectory('packages');
+      createPackagesDirectory(parentDir: packagesDir.parent);
+      processRunner = RecordingProcessRunner();
+      final PubspecCheckCommand command = PubspecCheckCommand(
+        packagesDir,
+        processRunner: processRunner,
+        platform: mockPlatform,
+      );
+
+      runner = CommandRunner<void>(
+          'pubspec_check_command', 'Test for pubspec_check_command');
+      runner.addCommand(command);
+    });
+
+    test('repository check works', () async {
+      final Directory packageDirectory =
+          createFakePackage('package', packagesDir);
+
+      packageDirectory.childFile('pubspec.yaml').writeAsStringSync('''
+${_headerSection('package')}
+${_environmentSection()}
+${_dependenciesSection()}
+''');
+
+      final List<String> output =
+          await runCapturingPrint(runner, <String>['pubspec-check']);
+
+      expect(
+        output,
+        containsAllInOrder(<Matcher>[
+          contains('Running for package...'),
           contains('No issues found!'),
         ]),
       );